feat: move ~/.factorio out of home
The wiki on this was confusing but i managed to figure this out.
You can set the user data directory with a config.ini file.
This file is by default located in `~/.factorio/config/config.ini`.
Since we want to move the entire `~/.factorio` directory we need to move
the config file as well.
The config file location can be specified with the command line
parameter `-c` or `--config`.
I added this to my launch options for Factorio on steam:
`-c /home/aiz/etc/factorio/config.ini`
And in the config.ini file I added this:
```
[path]
write-data=/home/aiz/var/share/factorio
```
I could then move the `~/.factorio` directory to `~/var/share/factorio`.
Despite the wiki using `.local/share/factorio` in it's example I
couldn't get relative paths to work.
Related issue: #1

Markdown in this commit might not work idk.