feat: start qbittorrent with systemd user unit
Starting qbittorrent with systemd is usefull because it means i don't
need to do anything about when moving to wayland.
It's also better for wayland since each wayland compositors need to have
their own init scripts.
So if I wanted to start qbittorrent in river and sway I would have to
do that in both their configuration files.
This solves that problem.