feat: add zoxide
I wanna try zoxide since I want faster navigation.
Due to muscle memory and the Z key being the Z key I've added a alias
for "cd" to "z"

Z might be easier to press on a staggered qwerty keyboard since you can
press it with your ring finger.
When comparing CD though the story changes.
on colemak-DH C and D are right next to each other on the bottom row and
you press them with the middle and index finger respectively.
This is a much faster and more fluid motion compared to qwerty.
I see "cd" as both faster and more comfortable to type compared to "z".

TL;DR learn ortholinear colemak-DH.