feat: source scripts in /etc/x11/xinit/xinitrc.d
Since the proper way to import DISPLAY & XAUTHORITY into dbus and
systemd is to source the /etc/X11/xinit/xinitrc.d/50-systemd-user.sh
file I've gone ahead and done that.

While this is nice I probably need to do it with river too and a
solution that works for both would be nice.