feat: add check-twitch.sh & starts it with systemd
I fixed some issues I had with my check-twitch.sh script and feel
comfortable tracking it now.
Instead of setting followed channels within the script it's now set via a universal environment variable.

followed channels is not something i want to track and making a config file is too much work for a script only I'm going to use.

Starting it with systemd means I don't need to do anything when
switching to wayland and stuff.

I don't really like that i need to use my username when setting the path
for the script and it's something i need to look into solving in the future.

The reason I use this script is because twitch stops sending
notifications if I never log into my account and because they banned me
for impersonation or some stupid stuff.
I don't think I ever interacted with someone using that account so I
don't want to make a new one.

I can't use twitchnotifier with the --interval flag when checking if a
channel is online so I needed to make a script to do that for me.